Renamed autoTriggerStrategy to holeFiller
