--- src/services/ghost/classic-auto-complete/GhostSuggestions.ts
@@ -1,5 +1,0 @@
-export interface FillInAtCursorSuggestion {
-	text: string
-	prefix: string
-	suffix: string
-}
