--- src/services/ghost/classic-auto-complete/ghostConstants.ts
@@ -1,4 +1,0 @@
-/**
- * Special marker used to indicate where completions should occur in the document
- */
-export const CURSOR_MARKER = "<<<AUTOCOMPLETE_HERE>>>"
